Update efa-dp-direct to v2
