+++ src/transport/net_efa_gda/efa-dp-direct/include/device/efa_cuda_dp.cuh
@@ -7,8 +7,25 @@
 #include <cstddef>
 #include <stdint.h>
 #include <cuda_runtime.h>
-#include "../host/efa_cuda_dp.h"
-#include "../common/efa_io_defs.h"
+
+#include "efa_cuda_dp_types.h"
+#include "efa_io_defs.h"
+
+enum efa_cuda_wc_opcode {
+	EFA_CUDA_WC_SEND,
+	EFA_CUDA_WC_RDMA_WRITE,
+	EFA_CUDA_WC_RDMA_READ,
+/*
+ * Set value of EFA_CUDA_WC_RECV so consumers can test if a completion is a
+ * receive by testing (opcode & EFA_CUDA_WC_RECV).
+ */
+	EFA_CUDA_WC_RECV                  = 1 << 7,
+	EFA_CUDA_WC_RECV_RDMA_WITH_IMM,
+};
+
+enum efa_cuda_processing_hint {
+	EFA_CUDA_PROCESSING_HINT_BURST_PPS_SENSITIVE = 1 << 0,
+};
 
 __device__ void *efa_cuda_cq_poll(efa_cuda_cq *cq, uint32_t position);
 
@@ -50,27 +67,20 @@ __device__ int efa_cuda_wr_set_inline_data(void *wr_buf, void *addr, size_t leng
 
 __device__ int efa_cuda_wr_set_sge(void *wr_buf, uint32_t lkey, uint64_t addr, uint32_t length);
 
+__device__ void efa_cuda_wr_set_processing_hints(void *wr_buf, uint32_t hints);
+
 __device__ void efa_cuda_flush_sq_wrs(efa_cuda_qp *qp);
 
 __device__ int efa_cuda_start_sq_batch(efa_cuda_qp *qp, int batch_size);
 
 __device__ int efa_cuda_sq_batch_place_wr(efa_cuda_qp *qp, int index_in_batch, void *wr_buf);
 
-__device__ int efa_cuda_post_recv_wr(efa_cuda_qp *qp, uint64_t addr, uint32_t length, uint32_t lkey);
+__device__ int efa_cuda_post_recv_wr(efa_cuda_qp *qp, uint16_t req_id, uint64_t addr, uint32_t length, uint32_t lkey);
 
 __device__ void efa_cuda_flush_rq_wrs(efa_cuda_qp *qp);
 
 __device__ bool efa_cuda_is_cq_compatible(efa_cuda_cq *cq);
 
 __device__ bool efa_cuda_is_qp_compatible(efa_cuda_qp *qp);
-
-namespace efa_cuda_dp {
-
-	struct efa_cuda_cq *create_cq(struct efa_cuda_cq_attrs *attrs, uint32_t inlen);
-	void destroy_cq(efa_cuda_cq *d_cq);
-	struct efa_cuda_qp *create_qp(struct efa_cuda_qp_attrs *attrs, uint32_t inlen);
-	void destroy_qp(efa_cuda_qp *d_qp);
-	int get_version(int *major, int *minor, int *subminor);
-}
 
 #endif
